+++ src/serverCore.js
@@ -31,15 +31,15 @@ class ServerCore {
 
   async saveServer(data_obj) {
     logger.debug('saveServer service called');
-    const { url, config } = data_obj;
+    const { url, namespace, config } = data_obj;
 
-    return await new Server(url, config).save();
+    return await new Server(url, namespace, config).save();
   }
 
   async getAllServers() {
     logger.debug('getAllServers service called');
 
-    return await Server.fetchAll(diagram_id);
+    return await Server.fetchAll();
   }
 
   async deleteServer(server_id) {